Experimenting with a 'Status' section
On one of my original designs for the site, there's a section called 'I am currently...' which I think is a neat idea. It's a personal insight, but also serves as a good indicator of how recently the site was last updated.

I can certainly update this more easily than adding new Journal entries.