Adding much more placeholder content
I think I need to revisit my designs for the 'portfolio' section of the site. It's feeling very bland at present and I have some ideas I'd like to play around with.

Also adding content to journal placeholder entries in an effort to spark of some inspiration and get writing