Finally adding in brand colours
Also removing middle column in the 3-column layout. It took me ages to realise what was making the home page feel so crowded. I never really intended for the main focus of the site to be a portfolio anyway so it makes sense to exclude 'work' from the home page.